code refactoring and api call optimization
