+++ src/components/category/ShowAllCategories.js
@@ -7,115 +7,112 @@ import coverImg from "../../utils/cover-img.jpg";
 import "./category.css";
 
 const ShowAllCategories = () => {
-	const dispatch = useDispatch();
-	const navigate = useNavigate();
+  const dispatch = useDispatch();
+  const navigate = useNavigate();
 
-	const category = useSelector((state) => state.category);
-	const city = useSelector((state) => state.city);
+  const category = useSelector((state) => state.category);
+  const city = useSelector((state) => state.city);
 
-	const getProductfromApi = async (ctg) => {
-		await api
-			.getProductbyFilter(
-				city.city.id,
-				city.city.latitude,
-				city.city.longitude,
-				{ category_id: ctg.id }
-			)
-			.then((response) => response.json())
-			.then((result) => {
-				if (result.status === 1) {
-					setMap(new Map(map.set(`category${ctg.id}`, result.total)));
-				}
-			});
-	};
+  const [map, setMap] = useState(new Map());
 
-	//fetch Category
-	const fetchCategory = () => {
-		api
-			.getCategory()
-			.then((response) => response.json())
-			.then((result) => {
-				if (result.status === 1) {
-					const dataToBeSorted = result.data;
-					//sorting of items lexographically..
-					const strAscending = [...dataToBeSorted].sort((a, b) =>
-						a.name > b.name ? 1 : -1
-					);
+  // const getProductfromApi = async (ctg) => {
+  //     console.log('showAllCategories');
+  // 	await api
+  // 		.getProductbyFilter(
+  // 			city.city.id,
+  // 			city.city.latitude,
+  // 			city.city.longitude,
+  // 			{ category_id: ctg.id }
+  // 		)
+  // 		.then((response) => response.json())
+  // 		.then((result) => {
+  // 			if (result.status === 1) {
+  // 				setMap(new Map(map.set(`category${ctg.id}`, result.total)));
+  // 			}
+  // 		});
+  // };
 
-					dispatch({ type: ActionTypes.SET_CATEGORY, payload: strAscending });
-				}
-			})
-			.catch((error) => console.log("error ", error));
-	};
+  //fetch Category
+  const fetchCategory = () => {
+    api
+      .getCategory()
+      .then((response) => response.json())
+      .then((result) => {
+        if (result.status === 1) {
+          const dataToBeSorted = result.data;
+          //sorting of items lexographically..
+          const strAscending = [...dataToBeSorted].sort((a, b) =>
+            a.name > b.name ? 1 : -1
+          );
 
-	useEffect(() => {
-		if (category.status === "loading" && category.category === null) {
-			fetchCategory();
-		} else if (category.status !== "loading" || city.city !== null) {
-			category.category.forEach((ctg) => {
-				getProductfromApi(ctg);
-			});
-		}
-	}, [category]);
+          dispatch({ type: ActionTypes.SET_CATEGORY, payload: strAscending });
+        }
+      })
+      .catch((error) => console.log("error ", error));
+  };
 
-	//categories and their product count map
-	const [map, setMap] = useState(new Map());
+  useEffect(() => {
+    if (category.status === "loading" && category.category === null) {
+      fetchCategory();
+    }
+  }, [category]);
 
-	const selectCategory = (category) => {
-		dispatch({ type: ActionTypes.SET_FILTER_CATEGORY, payload: category.id });
-		navigate("/products");
-	};
+  const selectCategory = (category) => {
+    dispatch({ type: ActionTypes.SET_FILTER_CATEGORY, payload: category.id });
+    navigate("/products");
+  };
 
-	return (
-		<section id="allcategories">
-			<div className="cover">
-				<img src={coverImg} className="img-fluid" alt="cover"></img>
-				<div className="page-heading">
-					<h5>Categories</h5>
-					<p>
-						<Link to={"/"}>home / </Link>
-						<span>Categories</span>
-					</p>
-				</div>
-			</div>
+  return (
+    <section id="allcategories">
+      <div className="cover">
+        <img src={coverImg} className="img-fluid" alt="cover"></img>
+        <div className="page-heading">
+          <h5>Categories</h5>
+          <p>
+            <Link to={"/"}>home / </Link>
+            <span>Categories</span>
+          </p>
+        </div>
+      </div>
 
-			<div className="container" style={{ padding: "30px 0" }}>
-				{category.status === "loading" ? (
-					<div className="d-flex justify-content-center">
-						<div className="spinner-border" role="status">
-							<span className="visually-hidden">Loading...</span>
-						</div>
-					</div>
-				) : (
-					<div className="row justify-content-center">
-						{category.category.map((ctg, index) => (
-							<div
-								className="col-md-2 col-sm-4 col-8  my-3 content"
-								key={index}
-							>
-								<div className="card">
-									<img className="card-img-top" src={ctg.image_url} alt="" />
-									<div
-										className="card-body"
-										style={{ cursor: "pointer" }}
-										onClick={() => selectCategory(ctg)}
-									>
-										<p>
-											{ctg.name} (
-											{map.get(`category${ctg.id}`) !== undefined
-												? map.get(`category${ctg.id}`)
-												: 0}
-											)
-										</p>
-									</div>
-								</div>
-							</div>
-						))}
-					</div>
-				)}
-			</div>
-		</section>
-	);
+      <div className="container" style={{ padding: "30px 0" }}>
+        {category.status === "loading" ? (
+          <div className="d-flex justify-content-center">
+            <div className="spinner-border" role="status">
+              <span className="visually-hidden">Loading...</span>
+            </div>
+          </div>
+        ) : (
+          <div className="row justify-content-center">
+            {category.category.map((ctg, index) => (
+              <div
+                className="col-md-2 col-sm-4 col-8  my-3 content"
+                key={index}
+              >
+                <div className="card">
+                  <div className="imageWrapper">
+                    <img className="card-img-top" src={ctg.image_url} alt="" />
+                  </div>
+                  <div
+                    className="card-body"
+                    style={{ cursor: "pointer" }}
+                    onClick={() => selectCategory(ctg)}
+                  >
+                    <p>
+                      {ctg.name}
+                      {/* (
+						{map.get(`category${ctg.id}`) !== undefined? map.get(`category${ctg.id}`): 0}
+						) */}
+                    </p>
+                  </div>
+                </div>
+              </div>
+            ))}
+          </div>
+        )}
+      </div>
+    </section>
+  );
 };
 
 export default ShowAllCategories;
