+++ src/components/homecontainer/HomeContainer.js
@@ -1,33 +1,23 @@
-import React, { useRef } from 'react'
-import { AiOutlineArrowRight } from 'react-icons/ai'
-import { FaChevronLeft, FaChevronRight } from 'react-icons/fa'
-import { Link } from 'react-router-dom'
-import Category from '../category/Category'
-import Slider from '../sliders/Slider'
-import './homecontainer.css'
+import React from "react";
+import Category from "../category/Category";
+import Slider from "../sliders/Slider";
+import "./homecontainer.css";
 
 const HomeContainer = () => {
-    
-    return (
+  return (
+    <section id="home" className="home-section container home-element section">
+      <div className="home-container row">
+        <div className="col-md-12 p-0 col-12">
+          <Slider />
+        </div>
+      </div>
+      <div className="category_section">
+        <div className="container">
+          <Category />
+        </div>
+      </div>
+    </section>
+  );
+};
 
-        // elementor-section-height-min-height elementor-section-items-stretch elementor-section-boxed elementor-section-height-default
-        <section id="home" className='home-section container home-element section'>
-            {/* Slider & Category */}
-            <div className='home-container row'>
-                <div className="col-md-12 p-0 col-12">
-                    <Slider />
-                </div>
-            </div>
-            <div className='category_section'>
-                <div className="container">
-
-                    <Category />
-
-                </div>
-            </div>
-        </section>
-
-    )
-}
-
-export default HomeContainer
+export default HomeContainer;
