+++ src/components/product/ProductDetails.js
@@ -33,6 +33,9 @@ import {
   incrementProduct,
 } from "../../services/cartService";
 import LoginUser from "../login/login-user";
+import RelateProducts from "./related-products";
+import RelateProduct from "./related-products";
+import Product from "./product";
 
 function SamplePrevArrow(props) {
   const { className, style, onClick } = props;
@@ -88,7 +91,6 @@ const ProductDetails = ({
   setProductTriggered = () => {},
 }) => {
   const dispatch = useDispatch();
-  const navigate = useNavigate();
   const cookies = new Cookies();
   const { slug } = useParams();
 
@@ -112,6 +114,7 @@ const ProductDetails = ({
   const [relatedProducts, setrelatedProducts] = useState(null);
   const [selectedProduct, setselectedProduct] = useState({});
   const [isLogin, setIsLogin] = useState(false);
+  const [isViewModal, setIsViewModal] = useState(false);
 
   const getCategoryDetails = (id) => {
     api
@@ -186,6 +189,7 @@ const ProductDetails = ({
 
   useEffect(() => {
     const findProductBySlug = async () => {
+      console.log("productDetails1");
       await api
         .getProductbyFilter(
           city.city.id,
@@ -213,6 +217,7 @@ const ProductDetails = ({
 
   useEffect(() => {
     if (Object.keys(productdata).length !== 0) {
+      console.log("productDetails2");
       api
         .getProductbyFilter(
           city.city.id,
@@ -250,6 +255,7 @@ const ProductDetails = ({
       removeSelectedProductId();
     };
   }, []);
+
   const settings = {
     infinite: false,
     slidesToShow: 5,
@@ -292,67 +298,6 @@ const ProductDetails = ({
         },
       },
     ],
-  };
-  const settings_subImage = {
-    infinite: false,
-    slidesToShow: 3,
-    initialSlide: 0,
-    // centerMargin: "10px",
-    margin: "20px",
-    prevArrow: (
-      <button
-        type="button"
-        className="slick-prev"
-        onClick={(e) => {
-          setmainimage(e.target.value);
-        }}
-      >
-        <FaChevronLeft size={30} className="prev-arrow" />
-      </button>
-    ),
-    nextArrow: (
-      <button
-        type="button"
-        className="slick-next"
-        onClick={(e) => {
-          setmainimage(e.target.value);
-        }}
-      >
-        <FaChevronRight color="#f7f7f7" size={30} className="next-arrow" />
-      </button>
-    ),
-    responsive: [
-      {
-        breakpoint: 1024,
-        settings: {
-          slidesToShow: 4,
-          slidesToScroll: 1,
-          infinite: true,
-        },
-      },
-      {
-        breakpoint: 768,
-        settings: {
-          slidesToShow: 3,
-          slidesToScroll: 1,
-        },
-      },
-    ],
-  };
-
-  //for product variants
-  const getProductSizeUnit = (variant) => {
-    if (Object.keys(productSize).length > 0) {
-      return productSize.map((psize) => {
-        if (
-          parseInt(psize.size) === parseInt(variant.measurement) &&
-          psize.short_code === variant.stock_unit_name
-        ) {
-          return psize.unit_id;
-        }
-      });
-    }
-    return <></>;
   };
 
   const getProductVariants = (product) => {
@@ -446,269 +391,6 @@ const ProductDetails = ({
       });
   };
 
-  //Add to favorite
-  const addToFavorite = async (product_id) => {
-    await api
-      .addToFavotite(cookies.get("jwt_token"), product_id)
-      .then((response) => response.json())
-      .then(async (result) => {
-        if (result.status === 1) {
-          toast.success(result.message);
-          await api
-            .getFavorite(
-              cookies.get("jwt_token"),
-              city.city.latitude,
-              city.city.longitude
-            )
-            .then((resp) => resp.json())
-            .then((res) => {
-              if (res.status === 1)
-                dispatch({ type: ActionTypes.SET_FAVORITE, payload: res });
-            });
-        } else {
-          toast.error(result.message);
-        }
-      });
-  };
-  const favorite = useSelector((state) => state.favorite);
-  const removefromFavorite = async (product_id) => {
-    await api
-      .removeFromFavorite(cookies.get("jwt_token"), product_id)
-      .then((response) => response.json())
-      .then(async (result) => {
-        if (result.status === 1) {
-          toast.success(result.message);
-          await api
-            .getFavorite(
-              cookies.get("jwt_token"),
-              city.city.latitude,
-              city.city.longitude
-            )
-            .then((resp) => resp.json())
-            .then((res) => {
-              if (res.status === 1)
-                dispatch({ type: ActionTypes.SET_FAVORITE, payload: res });
-              else dispatch({ type: ActionTypes.SET_FAVORITE, payload: null });
-            });
-        } else {
-          toast.error(result.message);
-        }
-      });
-  };
-
-  const addProductToCart1 = () => {
-    if (cookies.get("jwt_token") !== undefined) {
-      document
-        .getElementById(`Add-to-cart-productdetail`)
-        .classList.toggle("visually-hidden");
-      document
-        .getElementById(`input-cart-productdetail`)
-        .classList.toggle("visually-hidden");
-      document.getElementById(`input-productdetail`).innerHTML = 1;
-      addtoCart(
-        productdata.id,
-        JSON.parse(
-          document.getElementById(`select-product-variant-productdetail`).value
-        ).id,
-        document.getElementById(`input-productdetail`).innerHTML
-      );
-    } else {
-      const isAdded = addProductToCart(productdata);
-      if (isAdded) {
-        document
-          .getElementById(`Add-to-cart-productdetail`)
-          .classList.toggle("visually-hidden");
-        document
-          .getElementById(`input-cart-productdetail`)
-          .classList.toggle("visually-hidden");
-        document.getElementById(`input-productdetail`).innerHTML = 1;
-        setProductTriggered(!productTriggered);
-      }
-    }
-  };
-
-  const handleAddToCart = (index, related_product) => {
-    if (cookies.get("jwt_token") !== undefined) {
-      document
-        .getElementById(`Add-to-cart-section${index}`)
-        .classList.remove("active");
-      document
-        .getElementById(`input-cart-section${index}`)
-        .classList.add("active");
-      document.getElementById(`input-section${index}`).innerHTML = 1;
-      addtoCart(
-        related_product.id,
-        JSON.parse(
-          document.getElementById(`select-product${index}-variant-section`)
-            .value
-        ).id,
-        document.getElementById(`input-section${index}`).innerHTML
-      );
-    } else {
-      addProductToCart(related_product);
-      document
-        .getElementById(`Add-to-cart-section${index}`)
-        .classList.remove("active");
-      document
-        .getElementById(`input-cart-section${index}`)
-        .classList.add("active");
-      document.getElementById(`input-section${index}`).innerHTML = 1;
-      setProductTriggered(!productTriggered);
-    }
-  };
-
-  const handleDecrement = () => {
-    var val = parseInt(
-      document.getElementById(`input-productdetail`).innerHTML
-    );
-    if (cookies.get("jwt_token") !== undefined) {
-      if (val === 1) {
-        document.getElementById(`input-productdetail`).innerHTML = 0;
-        document
-          .getElementById(`Add-to-cart-productdetail`)
-          .classList.remove("visually-hidden");
-        document
-          .getElementById(`input-cart-productdetail`)
-          .classList.toggle("visually-hidden");
-        removefromCart(
-          productdata.id,
-          JSON.parse(
-            document.getElementById(`select-product-variant-productdetail`)
-              .value
-          ).id
-        );
-      } else {
-        document.getElementById(`input-productdetail`).innerHTML = val - 1;
-        addtoCart(
-          productdata.id,
-          JSON.parse(
-            document.getElementById(`select-product-variant-productdetail`)
-              .value
-          ).id,
-          document.getElementById(`input-productdetail`).innerHTML
-        );
-      }
-    } else {
-      const isDecremented = decrementProduct(productdata.id, productdata);
-      if (isDecremented) {
-        document.getElementById(`input-productdetail`).innerHTML = val - 1;
-      } else {
-        document.getElementById(`input-productdetail`).innerHTML = 0;
-        document
-          .getElementById(`Add-to-cart-productdetail`)
-          .classList.remove("visually-hidden");
-        document
-          .getElementById(`input-cart-productdetail`)
-          .classList.toggle("visually-hidden");
-      }
-      setProductTriggered(!productTriggered);
-    }
-  };
-
-  const handleDecrement1 = (related_product, index) => {
-    var val = parseInt(
-      document.getElementById(`input-section${index}`).innerHTML
-    );
-    if (cookies.get("jwt_token") !== undefined) {
-      if (val === 1) {
-        document.getElementById(`input-section${index}`).innerHTML = 0;
-        document
-          .getElementById(`input-cart-section${index}`)
-          .classList.remove("active");
-        document
-          .getElementById(`Add-to-cart-section${index}`)
-          .classList.add("active");
-        removefromCart(
-          related_product.id,
-          JSON.parse(
-            document.getElementById(`select-product${index}-variant-section`)
-              .value
-          ).id
-        );
-      } else {
-        document.getElementById(`input-section${index}`).innerHTML = val - 1;
-        addtoCart(
-          related_product.id,
-          JSON.parse(
-            document.getElementById(`select-product${index}-variant-section`)
-              .value
-          ).id,
-          document.getElementById(`input-section${index}`).innerHTML
-        );
-      }
-    } else {
-      const isDecremented = decrementProduct(
-        related_product.id,
-        related_product
-      );
-      if (isDecremented) {
-        document.getElementById(`input-section${index}`).innerHTML = val - 1;
-      } else {
-        document.getElementById(`input-section${index}`).innerHTML = 0;
-        document
-          .getElementById(`input-cart-section${index}`)
-          .classList.remove("active");
-        document
-          .getElementById(`Add-to-cart-section${index}`)
-          .classList.add("active");
-      }
-      setProductTriggered(!productTriggered);
-    }
-  };
-
-  const handleIncrement = () => {
-    var val = document.getElementById(`input-productdetail`).innerHTML;
-    if (cookies.get("jwt_token") !== undefined) {
-      if (val < productdata.total_allowed_quantity) {
-        document.getElementById(`input-productdetail`).innerHTML =
-          parseInt(val) + 1;
-        addtoCart(
-          productdata.id,
-          JSON.parse(
-            document.getElementById(`select-product-variant-productdetail`)
-              .value
-          ).id,
-          document.getElementById(`input-productdetail`).innerHTML
-        );
-      }
-    } else {
-      const isIncremented = incrementProduct(productdata.id, productdata);
-      if (isIncremented) {
-        document.getElementById(`input-productdetail`).innerHTML =
-          parseInt(val) + 1;
-      }
-      setProductTriggered(!productTriggered);
-    }
-  };
-
-  const handleIncrement1 = (related_product, index) => {
-    var val = document.getElementById(`input-section${index}`).innerHTML;
-    if (cookies.get("jwt_token") !== undefined) {
-      if (val < related_product.total_allowed_quantity) {
-        document.getElementById(`input-section${index}`).innerHTML =
-          parseInt(val) + 1;
-        addtoCart(
-          related_product.id,
-          JSON.parse(
-            document.getElementById(`select-product${index}-variant-section`)
-              .value
-          ).id,
-          document.getElementById(`input-section${index}`).innerHTML
-        );
-      }
-    } else {
-      const isIncremented = incrementProduct(
-        related_product.id,
-        related_product
-      );
-      if (isIncremented) {
-        document.getElementById(`input-section${index}`).innerHTML =
-          parseInt(val) + 1;
-      }
-      setProductTriggered(!productTriggered);
-    }
-  };
-
   return (
     <div className="product-details-view">
       <div className="container" style={{ gap: "20px" }}>
@@ -722,242 +404,18 @@ const ProductDetails = ({
               </div>
             </div>
           ) : (
-            <div className="row body-wrapper ">
-              <div className="col-xl-4 col-lg-6 col-md-12 col-12">
-                <div className="image-wrapper ">
-                  <div className="main-image col-12 border">
-                    <img
-                      src={mainimage}
-                      alt="main-product"
-                      className="col-12"
-                      style={{ width: "85%" }}
-                    />
-                  </div>
-
-                  <div className="sub-images-container">
-                    {images.length >= 4 ? (
-                      <>
-                        <Slider {...settings_subImage}>
-                          {images.map((image, index) => (
-                            <div key={index}>
-                              <div
-                                className={`sub-image border ${
-                                  mainimage === image ? "active" : ""
-                                }`}
-                              >
-                                <img
-                                  src={image}
-                                  className="col-12"
-                                  alt="product"
-                                  onClick={() => {
-                                    setmainimage(image);
-                                  }}
-                                ></img>
-                              </div>
-                            </div>
-                          ))}
-                        </Slider>
-                      </>
-                    ) : (
-                      <>
-                        {images.map((image, index) => (
-                          <div
-                            key={index}
-                            className={`sub-image border ${
-                              mainimage === image ? "active" : ""
-                            }`}
-                          >
-                            <img
-                              src={image}
-                              className="col-12 "
-                              alt="product"
-                              onClick={() => {
-                                setmainimage(image);
-                              }}
-                            ></img>
-                          </div>
-                        ))}
-                      </>
-                    )}
-                  </div>
-                </div>
-              </div>
-              <div className="col-xl-8 col-lg-6 col-md-12 col-12">
-                <div className="detail-wrapper">
-                  <div className="top-section">
-                    <p className="product_name">{productdata.name}</p>
-                    {Object.keys(productbrand).length === 0 ? null : (
-                      <div className="product-brand">
-                        <span className="brand-title">Brand:</span>
-                        <span className="brand-name">{productbrand.name}</span>
-                      </div>
-                    )}
-                    <div className="d-flex flex-row gap-2 align-items-center my-1">
-                      <span
-                        className="price green-text"
-                        id={`price-productdetail`}
-                      >
-                        <FaRupeeSign fill="var(--secondary-color)" />
-                        {parseFloat(productdata.variants[0].price)}{" "}
-                      </span>{" "}
-                      <div
-                        className="not-price gray-text"
-                        style={{ textDecoration: "line-through" }}
-                      >
-                        <FaRupeeSign
-                          fill="var(--text-color)"
-                          textDecoration="line-through"
-                        />
-                        {parseFloat(productdata.variants[0].price) +
-                          productdata.variants[0].price * 0.13}
-                      </div>
-                    </div>
-                  </div>
-                  <div className="bottom-section">
-                    <p>Product Variants</p>
-                    <div className="d-flex gap-3 bottom-section-content">
-                      <select
-                        id={`select-product-variant-productdetail`}
-                        onChange={(e) => {
-                          document.getElementById(
-                            `price-productdetail`
-                          ).innerHTML = parseFloat(
-                            JSON.parse(e.target.value).price
-                          );
-
-                          if (
-                            document
-                              .getElementById(`input-cart-productdetail`)
-                              .classList.contains("active")
-                          ) {
-                            document
-                              .getElementById(`input-cart-productdetail`)
-                              .classList.remove("active");
-                            document
-                              .getElementById(`Add-to-cart-productdetail`)
-                              .classList.add("active");
-                          }
-                        }}
-                        defaultValue={JSON.stringify(productdata.variants[0])}
-                      >
-                        {getProductVariants(productdata)}
-                      </select>
-
-                      <button
-                        type="button"
-                        id={`Add-to-cart-productdetail`}
-                        className="add-to-cart active"
-                        onClick={addProductToCart1}
-                      >
-                        Add to Cart
-                      </button>
-                      <div
-                        id={`input-cart-productdetail`}
-                        className="input-to-cart visually-hidden"
-                      >
-                        <button
-                          type="button"
-                          className="wishlist-button"
-                          onClick={() => {
-                            handleDecrement();
-                          }}
-                        >
-                          <BiMinus fill="#fff" />
-                        </button>
-                        <span id={`input-productdetail`}></span>
-                        <button
-                          type="button"
-                          className="wishlist-button"
-                          onClick={() => {
-                            handleIncrement();
-                          }}
-                        >
-                          <BsPlus fill="#fff" />{" "}
-                        </button>
-                      </div>
-                      {/* <button type='button' className='wishlist-product' onClick={() => addToFavorite(productdata.id)}><BsHeart /></button> */}
-                      {favorite.favorite &&
-                      favorite.favorite.data.some(
-                        (element) => element.id === productdata.id
-                      ) ? (
-                        <button
-                          type="button"
-                          className="wishlist-product"
-                          onClick={() => removefromFavorite(productdata.id)}
-                        >
-                          <BsHeartFill fill="green" />
-                        </button>
-                      ) : (
-                        <button
-                          key={productdata.id}
-                          type="button"
-                          className="wishlist-product"
-                          onClick={() => addToFavorite(productdata.id)}
-                        >
-                          <BsHeart />
-                        </button>
-                      )}
-                    </div>
-                    <div className="product-overview">
-                      <div className="product-seller">
-                        {/* <span className='seller-title'>Sold By:</span>
-                                                    <span className='seller-name'>{productdata.seller_name} </span> */}
-                      </div>
-
-                      {productdata.tags !== "" ? (
-                        <div className="product-tags">
-                          <span className="tag-title">Product Tags:</span>
-                          <span className="tag-name">{productdata.tags} </span>
-                        </div>
-                      ) : (
-                        ""
-                      )}
-                    </div>
-                    <div className="share-product-container">
-                      <span>Share product:</span>
-
-                      <ul className="share-product">
-                        <li className="share-product-icon">
-                          <WhatsappShareButton
-                            url={`https://chhayakart.com/product/${productdata.slug}`}
-                          >
-                            <WhatsappIcon size={32} round={true} />{" "}
-                          </WhatsappShareButton>
-                        </li>
-                        <li className="share-product-icon">
-                          <TelegramShareButton
-                            url={`https://chhayakart.com/product/${productdata.slug}`}
-                          >
-                            <TelegramIcon size={32} round={true} />{" "}
-                          </TelegramShareButton>
-                        </li>
-                        <li className="share-product-icon">
-                          <FacebookShareButton
-                            url={`https://chhayakart.com/product/${productdata.slug}`}
-                          >
-                            <FacebookIcon size={32} round={true} />{" "}
-                          </FacebookShareButton>
-                        </li>
-                        <li className="share-product-icon">
-                          <button
-                            type="button"
-                            onClick={() => {
-                              navigator.clipboard.writeText(
-                                `https://chhayakart.com/product/${productdata.slug}`
-                              );
-                              toast.success("Copied Succesfully!!");
-                            }}
-                          >
-                            {" "}
-                            <BiLink size={30} />
-                          </button>
-                        </li>
-                      </ul>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
+            <Product
+              images={images}
+              mainimage={mainimage}
+              productbrand={productbrand}
+              setmainimage={setmainimage}
+              addtoCart={addtoCart}
+              productdata={productdata}
+              productTriggered={productTriggered}
+              setProductTriggered={setProductTriggered}
+              removefromCart={removefromCart}
+              getProductVariants={getProductVariants}
+            />
           )}
         </div>
 
@@ -985,234 +443,18 @@ const ProductDetails = ({
                   <Slider {...settings}>
                     {relatedProducts.map((related_product, index) => (
                       <div className="col-md-3 col-lg-4">
-                        <div className="product-card">
-                          <div className="image-container">
-                            <span
-                              className="border border-light rounded-circle p-2 px-3"
-                              id="aiEye"
-                              onClick={() => {
-                                setselectedProduct(related_product);
-                              }}
-                              data-bs-toggle="modal"
-                              data-bs-target="#quickviewModal"
-                            >
-                              <AiOutlineEye
-                                onClick={() => {
-                                  setselectedProduct(related_product);
-                                }}
-                                data-bs-toggle="modal"
-                                data-bs-target="#quickviewModal"
-                              />
-                            </span>
-                            <img
-                              src={related_product.image_url}
-                              alt={related_product.slug}
-                              className="card-img-top"
-                              onClick={() => {
-                                dispatch({
-                                  type: ActionTypes.SET_SELECTED_PRODUCT,
-                                  payload: related_product.id,
-                                });
-                                setSelectedProductId(related_product.id);
-                                navigate("/product/" + related_product.id);
-                                window.scrollTo(0, 0);
-                              }}
-                            />
-                          </div>
-
-                          <div
-                            className="card-body product-card-body p-3"
-                            onClick={() => {
-                              dispatch({
-                                type: ActionTypes.SET_SELECTED_PRODUCT,
-                                payload: related_product.id,
-                              });
-                              setSelectedProductId(related_product.id);
-                            }}
-                          >
-                            <h3>{related_product.name}</h3>
-                            <div className="price">
-                              <span
-                                id={`price${index}-section`}
-                                className="d-flex align-items-center"
-                              >
-                                <p id="fa-rupee" className="m-0">
-                                  <FaRupeeSign fill="var(--secondary-color)" />
-                                </p>
-                                {related_product.variants[0].price}{" "}
-                              </span>
-                              <span
-                                id={`price${index}-section`}
-                                className="d-flex align items-center"
-                                style={{ textDecoration: "line-through" }}
-                              >
-                                <p id="fa-rupee" className="m-0">
-                                  <FaRupeeSign fill="var(--secondary-color)" />
-                                </p>{" "}
-                                {parseFloat(related_product.variants[0].price) +
-                                  related_product.variants[0].price * 0.13}
-                              </span>
-                            </div>
-                            <div className="product_varients_drop">
-                              {related_product.variants.length > 1 ? (
-                                <>
-                                  <select
-                                    style={{ fontSize: "8px !important" }}
-                                    className="form-select variant_selection select-arrow"
-                                    id={`select-product${index}-variant-section`}
-                                    onChange={(e) => {
-                                      document.getElementById(
-                                        `price${index}-section`
-                                      ).innerHTML = parseFloat(
-                                        JSON.parse(e.target.value).price
-                                      );
-
-                                      if (
-                                        document
-                                          .getElementById(
-                                            `input-cart-section${index}`
-                                          )
-                                          .classList.contains("active")
-                                      ) {
-                                        document
-                                          .getElementById(
-                                            `input-cart-section${index}`
-                                          )
-                                          .classList.remove("active");
-                                        document
-                                          .getElementById(
-                                            `Add-to-cart-section${index}`
-                                          )
-                                          .classList.add("active");
-                                      }
-                                    }}
-                                    defaultValue={JSON.stringify(
-                                      related_product.variants[0]
-                                    )}
-                                  >
-                                    {getProductVariants(related_product)}
-                                  </select>
-                                </>
-                              ) : (
-                                <>
-                                  <input
-                                    type="hidden"
-                                    name=""
-                                    id={`select-product${index}-variant-section`}
-                                    value={JSON.stringify(
-                                      related_product.variants[0]
-                                    )}
-                                  />
-                                  {/* <span className='variant_value select-arrow' id=''>{product.variants[0].measurement + " " + product.variants[0].stock_unit_name}
-																	</span> */}
-                                  <span
-                                    className="variant_value select-arrow"
-                                    id=""
-                                  >
-                                    {
-                                      related_product.variants[0]
-                                        .stock_unit_name
-                                    }
-                                  </span>
-                                </>
-                              )}
-                            </div>
-                          </div>
-
-                          <div className="d-flex flex-row border-top product-card-footer">
-                            <div
-                              className="border-end"
-                              style={{ flexGrow: "1" }}
-                            >
-                              <button
-                                type="button"
-                                id={`Add-to-cart-section${index}`}
-                                className="w-100 h-100 add-to-cart active"
-                                onClick={() => {
-                                  handleAddToCart(index, related_product);
-                                }}
-                              >
-                                add to cart
-                              </button>
-
-                              <div
-                                id={`input-cart-section${index}`}
-                                className="w-100 h-100 input-to-cart"
-                              >
-                                <button
-                                  type="button"
-                                  onClick={() => {
-                                    handleDecrement1(related_product, index);
-                                  }}
-                                >
-                                  <BiMinus />
-                                </button>
-                                <span id={`input-section${index}`}></span>
-                                <button
-                                  type="button"
-                                  onClick={() => {
-                                    handleIncrement1(related_product, index);
-                                  }}
-                                >
-                                  <BsPlus />{" "}
-                                </button>
-                              </div>
-                            </div>
-
-                            <div className="dropup share">
-                              <button
-                                type="button"
-                                className="w-100 h-100 "
-                                data-bs-toggle="dropdown"
-                                aria-expanded="false"
-                              >
-                                <BsShare />
-                              </button>
-
-                              <ul className="dropdown-menu">
-                                <li className="dropDownLi">
-                                  <WhatsappShareButton
-                                    url={`https://chhayakart.com/product/${related_product.slug}`}
-                                  >
-                                    <WhatsappIcon size={32} round={true} />{" "}
-                                    <span>WhatsApp</span>
-                                  </WhatsappShareButton>
-                                </li>
-                                <li className="dropDownLi">
-                                  <TelegramShareButton
-                                    url={`https://chhayakart.com/product/${related_product.slug}`}
-                                  >
-                                    <TelegramIcon size={32} round={true} />{" "}
-                                    <span>Telegram</span>
-                                  </TelegramShareButton>
-                                </li>
-                                <li className="dropDownLi">
-                                  <FacebookShareButton
-                                    url={`https://chhayakart.com/product/${related_product.slug}`}
-                                  >
-                                    <FacebookIcon size={32} round={true} />{" "}
-                                    <span>Facebook</span>
-                                  </FacebookShareButton>
-                                </li>
-                                <li>
-                                  <button
-                                    type="button"
-                                    onClick={() => {
-                                      navigator.clipboard.writeText(
-                                        `https://chhayakart.com/product/${related_product.slug}`
-                                      );
-                                      toast.success("Copied Succesfully!!");
-                                    }}
-                                    className="react-share__ShareButton"
-                                  >
-                                    {" "}
-                                    <BiLink size={30} /> <span>Copy Link</span>
-                                  </button>
-                                </li>
-                              </ul>
-                            </div>
-                          </div>
-                        </div>
+                        <RelateProduct
+                          index={index}
+                          related_product={related_product}
+                          productTriggered={productTriggered}
+                          setIsViewModal={setIsViewModal}
+                          setselectedProduct={setselectedProduct}
+                          setSelectedProductId={setSelectedProductId}
+                          setProductTriggered={setProductTriggered}
+                          getProductVariants={getProductVariants}
+                          removefromCart={removefromCart}
+                          addtoCart={addtoCart}
+                        />
                       </div>
                     ))}
                   </Slider>
@@ -1220,12 +462,16 @@ const ProductDetails = ({
               )}
             </div>
           </div>
-          <QuickViewModal
-            selectedProduct={selectedProduct}
-            setselectedProduct={setselectedProduct}
-            productTriggered={productTriggered}
-            setProductTriggered={setProductTriggered}
-          />
+          {isViewModal && (
+            <QuickViewModal
+              selectedProduct={selectedProduct}
+              setselectedProduct={setselectedProduct}
+              productTriggered={productTriggered}
+              setProductTriggered={setProductTriggered}
+              isOpenModal={isViewModal}
+              setIsOpenModal={setIsViewModal}
+            />
+          )}
         </div>
 
         {isLogin && (
