+++ src/components/product/product.css
@@ -4,10 +4,6 @@
     padding: 0 !important;
     padding-left: -2px;
   }
-
-  .productDescription{
-    text-transform: uppercase;
-  }
   
   .css-b62m3t-container {
     width: 100%;
@@ -142,13 +138,12 @@
   }
   
   .product_title_content a {
-    width: 85px;
     background: rgba(81, 189, 136, 0.2);
     border: 1px solid var(--secondary-color);
     border-radius: 8px;
     text-transform: capitalize;
     font-size: 16px;
-    padding: 8px;
+    padding: 8px 16px;
     text-decoration: none;
     color: var(--secondary-color);
     text-align: center;
@@ -169,6 +164,7 @@
     border-radius: 8px;
     border: var(--bs-border-width) var(--bs-border-style) var(--bs-border-color) !important;
     overflow: hidden;
+    margin: 6px;
   }
   
   .product-card:hover {
@@ -207,10 +203,15 @@
     filter: invert(1);
   }
   
-  .product-card .image-container img {
-    height: auto;
+  .product-card .image-container .imageWrapper {
+    height: 180px;
+    width: auto;
+  }
+
+  .product-card .image-container .imageWrapper img{
+    object-fit: cover;
+    height: 100%;
     width: 100%;
-    /* padding: 15px; */
   }
   
   .product-card .product-card-body {
@@ -323,20 +324,24 @@
     display: none;
   }
   
-  .product-card .product-card-footer .add-to-cart.active,
-  .product-card .product-card-footer .input-to-cart.active {
+  .product-card .product-card-footer .add-to-cart,
+  .product-card .product-card-footer .input-to-cart {
     display: block;
     text-transform: uppercase;
   }
+
+  .cartCout{
+    width: 100%;
+  }
   
-  .product-card .product-card-footer .input-to-cart.active {
+  .product-card .product-card-footer .input-to-cart {
     background: var(--secondary-color);
     display: flex;
     justify-content: space-between;
     align-items: center;
   }
   
-  .product-card .product-card-footer .input-to-cart.active * {
+  .product-card .product-card-footer .input-to-cart * {
     color: #fff;
   }
   
@@ -455,7 +460,7 @@
     padding: 0 5px;
   }
   
-  #productlist .row .filter-view .icon.active {
+  #productlist .row .filter-view .icon {
     background: var(--secondary-color);
   }
   
@@ -729,7 +734,7 @@
     }
   }
   
-  .product-details-view .image-wrapper .sub-images-container .sub-image.active {
+  .product-details-view .image-wrapper .sub-images-container .sub-image {
     border: 1px solid var(--secondary-color);
     border-radius: 8px;
   }
@@ -999,7 +1004,7 @@
   .product-details-view
     .detail-wrapper
     .bottom-section
-    .input-to-cart.active
+    .input-to-cart
     span {
     background: #a7eeccc9;
     color: var(--secondary-color);
@@ -1174,25 +1179,4 @@
   #eye {
     border-radius: 50%;
     margin: -5px;
-  }
-
-  .rowWrapper{
-    display: flex;
-    gap: 12px;
-    align-items: center;
-    justify-content: space-between;
-  }
-
-  .product_description{
-    text-align: left;
-  }
-
-  @media only screen and (max-width: 575px){
-    .product_title_content p{
-      font-size: 14px;
-    }
-    
-    .product_title_content a{
-      font-size: 14px;
-    }
   }
