+++ src/components/shared/ck-modal.css
@@ -1,5 +1,6 @@
 .dkModal {
   z-index: 10009;
+  width: 100px;
 }
 
 .dkModal .close {
@@ -20,6 +21,11 @@
   font-weight: 700;
   line-height: 1;
   background: white;
+}
+
+.modal-dialog{
+  min-width: 80vw !important;
+  margin: 0 auto;
 }
 
 @media only screen and (max-width: 575px){
